+++ scripts/integrations/dae.js
@@ -2,6 +2,7 @@ export function addDAEFlags() {
     let crFlags = Object.keys(CONFIG.DND5E.conditionTypes).map(i => 'flags.chris-premades.CR.' + i);
     let cvFlags = Object.keys(CONFIG.DND5E.conditionTypes).map(i => 'flags.chris-premades.CV.' + i);
     DAE.addAutoFields(crFlags.concat(cvFlags));
+    DAE.addAutoFields(['flags.chris-premades.skill.grappleInitiate', 'flags.chris-premades.skill.grappleEscape']);
 }
 export function colorizeDAETitleBarButton(app, [elem], options) {
     let headerButton = elem.closest('.window-app').querySelector('a.open-item-effect');