+++ CHANGELOG.md
@@ -1,8 +1,14 @@
+# 0.9.62 Release:
+## New Content:
+- Plasmoid: Amorphous
+## Bug Fixes:
+- Skill Patching
+## Update Notes:
+- Support for D&D system version 3.0.0 will be in a future update. Do not update your system yet.
+  
 # 0.9.61 Release:
 ## Bug Fixes:
 - Stone's Endurance
-## Update Notes:
-- Support for D&D system version 3.0.0 will be in a future update. Do not update your system yet.
   
 # 0.9.60 Release:
 ## New Content: